requirements.txt
pandas
numpy
scikit-learn
joblib
shap
mlflow
fastapi
uvicorn
pytest
pyyaml